refactor(servers): 拆分 server-list god-component（1103→557）
server-list.tsx 1103 行 6 职责拆分，主组件回落 557 行纯编排：
- useSpeedTest / useServerFilter 两 hook（测速编排+事件订阅 / 搜索筛选排序）
- ServerCard / ServerRow / ServerActions 三展示组件
- server-list-helpers 纯函数+类型模块
行为字节级等价，对外 ServerListProps 接口零变更；不引状态库/context（避免过度工程）。
